fix : make ListExprOptiontype class and fix trailing comma
