+++ src/ai/backend/client/cli/params.py
@@ -1,5 +1,6 @@
 import json
 import re
+from decimal import Decimal
 from typing import (
     Any,
     Optional,
@@ -85,3 +86,55 @@ class JSONParamType(click.ParamType):
         except json.JSONDecodeError:
             self.fail(f"cannot parse {value!r} as JSON", param, ctx)
         return value
+
+
+def drange(start: Decimal, stop: Decimal, num: int):
+    """
+    A simplified version of numpy.linspace with default options
+    """
+    delta = stop - start
+    step = delta / (num - 1)
+    yield from (start + step * Decimal(tick) for tick in range(0, num))
+
+
+class RangeExprOptionType(click.ParamType):
+    """
+    Accepts a range expression which generates a range of values for a variable.
+
+    Linear space range: "linspace:1,2,10" (start, stop, num) as in numpy.linspace
+    Pythonic range: "range:1,10,2" (start, stop[, step]) as in Python's range
+    Case range: "case:a,b,c" (comma-separated strings)
+    """
+    _rx_range_key = re.compile(r'^[a-zA-Z_][a-zA-Z0-9_]*$')
+    name = 'Range Expression'
+
+    def convert(self, arg, param, ctx):
+        key, value = arg.split('=', maxsplit=1)
+        assert self._rx_range_key.match(key), 'The key must be a valid slug string.'
+        try:
+            if value.startswith('case:'):
+                return key, value[5:].split(',')
+            elif value.startswith('linspace:'):
+                start, stop, num = value[9:].split(',')
+                return key, tuple(drange(Decimal(start), Decimal(stop), int(num)))
+            elif value.startswith('range:'):
+                range_args = map(int, value[6:].split(','))
+                return key, tuple(range(*range_args))
+            else:
+                self.fail('Unrecognized range expression type', param, ctx)
+        except ValueError as e:
+            self.fail(str(e), param, ctx)
+
+
+class ListExprOptionType(click.ParamType):
+
+    name = 'List Expression'
+
+    def convert(self, arg, param, ctx):
+        try:
+            if isinstance(arg, int):
+                return arg.split(',')
+            elif isinstance(arg, str):
+                return arg.split(',')
+        except ValueError as e:
+            self.fail(repr(e), param, ctx)
